==> guess_model/8_human_refine.R <==
# Title     : Script to refine guess labels
# Objective : To apply a few steps to refine 
#             the guess labels generated by random forest.
# Created by: Lei Song
# Created on: 03/23/21

####################### 
##  Step 1: Setting  ##
#######################
message('Step 1: Setting')

library(glue)
library(here)
library(terra)
library(ramify)
library(sf)
library(rgrass7)
library(dplyr)
library(stringr)
library(tidyverse)

###############################
###  Step 2: Refine labels  ###
###############################
message('Step 2: Refine labels')

refine_scores <- function(tile_sf,
                          roads,
                          waterbodies,
                          buildings,
                          guess_dir,
                          refine_dir,
                          reduce_class,
                          reduce_ratio){
    tile_id <- tile_sf$tile
    index <- tile_sf$index
    scores <- rast(
        here(glue('{guess_dir}/scores_{tile_id}_{index}.tif')))
    
    reduce_class <- paste0('.pred_', reduce_class)
    for (n in 1:length(reduce_class)){
        scores[[reduce_class[n]]] <- 
            scores[[reduce_class[n]]] * reduce_ratio[n]
    }
    pred <- values(scores); pred[is.na(pred)] <- 0
    pred <- argmax(pred)
    classes <- scores[[1]]
    values(classes) <- pred
    rm(scores, pred); gc()
    
    # Crop vectors
    crs_mer <- crs(classes)
    road <- roads %>% 
        st_cast('MULTILINESTRING') %>% 
        st_transform(crs = crs_mer)
    waterbody <- waterbodies %>% 
        st_cast('MULTIPOLYGON') %>% 
        st_transform(crs = crs_mer)
    building <- buildings %>% 
        st_cast('MULTIPOLYGON') %>% 
        st_transform(crs = crs_mer)
    
    # Add vectors
    # Set up GRASS GIS
    message('------Add OSM layers')
    crs_mer <- terra::crs(classes, proj = T)
    gisBase <- '/Applications/GRASS-7.8.app/Contents/Resources'
    initGRASS(gisBase = gisBase,
              home = tempdir(),
              gisDbase = tempdir(),  
              mapset = 'PERMANENT', 
              location = glue('segments_{tile_id}_{index}'), 
              override = TRUE)
    execGRASS("g.proj", flags = "c", 
              proj4 = crs_mer)
    execGRASS('r.in.gdal', flags = c("o", "overwrite"),
              input = file.path(
                  guess_dir, 
                  glue('scores_{tile_id}_{index}.tif')),
              band = 1,
              output = "score")
    execGRASS("g.region", raster = "score")
    
    use_sf()
    if (nrow(road) > 0){
        writeVECT(road, 'roads', v.in.ogr_flags = 'overwrite')
        execGRASS('v.to.rast', flags = c("overwrite"),
                  parameters = list(input = 'roads', 
                                    output = 'roads',
                                    use = 'val',
                                    value = 3))
        roads_path <- tempfile()
        execGRASS('r.out.gdal', flags = c("m", "overwrite"),
                  output = roads_path,
                  input = "roads")
        fill_roads <- rast(roads_path)
    } else {
        fill_roads <- NULL
    }; rm(road)
    
    if (nrow(waterbody) > 0){
        writeVECT(waterbody, 'waterbodies', v.in.ogr_flags = 'overwrite')
        execGRASS('v.to.rast', flags = c("overwrite"),
                  parameters = list(input = 'waterbodies', 
                                    output = 'waterbodies',
                                    use = 'val',
                                    value = 5))
        waterbody_path <- tempfile()
        execGRASS('r.out.gdal', flags = c("m", "overwrite"),
                  output = waterbody_path,
                  input = "waterbodies")
        fill_waterbodies <- rast(waterbody_path)
    } else {
        fill_waterbodies <- NULL
    }; rm(waterbody)
    
    if (nrow(building) > 0){
        writeVECT(building, 'buildings', v.in.ogr_flags = 'overwrite')
        execGRASS('v.to.rast', flags = c("overwrite"),
                  parameters = list(input = 'buildings', 
                                    output = 'buildings',
                                    use = 'val',
                                    value = 6))
        building_path <- tempfile()
        execGRASS('r.out.gdal', flags = c("m", "overwrite"),
                  output = building_path,
                  input = "buildings")
        fill_buildings <- rast(building_path)
    } else {
        fill_buildings <- NULL
    }; rm(building)
    
    # Gather results
    fills <- compact(list(roads = fill_roads,
                          buildings = fill_buildings,
                          waterbodies = fill_waterbodies))
    rm(fill_roads, fill_buildings, fill_waterbodies)
    
    # Replace values
    if (length(fills) > 0){
        for (i in 1:length(fills)){
            mask_vct <- fills[[i]]
            mask_vct <- is.na(mask_vct)
            classes <- classes * mask_vct
            classes <- cover(classes, fills[[i]], values = 0)
            rm(mask_vct)
        }
    }
    
    # Save out
    message('------Save out')
    names(classes) <- 'label'
    writeRaster(classes, 
                file.path(refine_dir, 
                          glue('{tile_id}_{index}_label.tif')),
                overwrite = T,
                wopt = list(datatype = 'INT1U',
                            gdal=c("COMPRESS=LZW")))
}

# Define function to refine the labels
## Step 2-0: Reduce the weight of some classes.
## Step 2-1: Apply some hand draw masks
# Classes:
# 1: cropland
# 2: forest
# 3: grassland
# 4: shrubland
# 6: water
# 8: Urban
# 9: bareland

# Read vectors
tiles <- read_sf('results/north/catalog_sample_tiles.geojson')
tiles_big <- read_sf('data/geoms/tiles_nicfi_north.geojson')
roads_all <- read_sf(here('data/osm/roads.geojson'))
roads_all <- st_join(roads_all, tiles_big)
waterbodies_all <- read_sf(here('data/osm/waterbodies.geojson'))
waterbodies_all <- st_join(waterbodies_all, tiles_big)
buildings_all <- read_sf(here('data/osm/buildings.geojson'))
buildings_all <- st_join(buildings_all, tiles_big)

guess_dir <- here('results/north/guess_labels')
refine_dir <- here('results/north/refine_labels')
if (!dir.exists(refine_dir)) dir.create(refine_dir)

# Change the parameters correspondingly
tile_id <- '1219-1017'
tile_index <- '58'
convert <- T
convert_class <- c(8)
to_class <- c(6)
touch_scores <- F
big_road <- T
apply_mask <- F
apply_line_mask <- F
reduce_class <- c(1)
reduce_ratio <- c(0)
reduce_class <- c(reduce_class, 8)
reduce_ratio <- c(reduce_ratio, 0)

# Automatic fix
if (touch_scores){
    tile_sf <- tiles %>% 
        filter(tile == tile_id & index == tile_index)
    roads <- roads_all %>% 
        filter(tile == tile_id)
    if (big_road){
        roads <- roads %>%
            filter(fclass %in% c('primary', 'secondary', 'tertiary',
                                 'trunk', 'primary_link', 'secondary_link',
                                 'tertiary_link', 'rail'))
    }
    waterbodies <- waterbodies_all %>% 
        filter(tile == tile_id)
    buildings <- buildings_all %>% 
        filter(tile == tile_id)
    
    refine_scores(tile_sf,
                  roads,
                  waterbodies,
                  buildings,
                  guess_dir,
                  refine_dir,
                  reduce_class,
                  reduce_ratio)
} 
if (convert) {
    classes <- rast(
        here(glue('{guess_dir}/guess_{tile_id}_{tile_index}.tif')))
    
    if (length(convert_class) > 0){
        for (n in 1:length(convert_class)){
            classes[classes == convert_class[n]] <- 
                to_class[n]
        }
    }
    
    # Save out
    names(classes) <- 'label'
    writeRaster(classes, 
                file.path(refine_dir, 
                          glue('{tile_id}_{tile_index}_label.tif')),
                overwrite = T,
                wopt = list(datatype = 'INT1U',
                            gdal=c("COMPRESS=LZW")))
}

# Manual fix
if (apply_mask){
    mask_vct <- read_sf(here('results/north/mask_vct.geojson'))
    tile_sf <- tiles %>% 
        filter(tile == tile_id & index == tile_index) %>% 
        st_buffer(0.01)
    
    classes <- rast(
        here(glue('{refine_dir}/{tile_id}_{tile_index}_label.tif')))
    
    crs_mer <- crs(classes)
    mask_vct <- mask_vct %>% 
        st_make_valid() %>% 
        st_intersection(tile_sf) %>%
        st_transform(crs = crs_mer) %>% 
        dplyr::select(landcover) %>% 
        st_cast('MULTIPOLYGON')
    
    crs_mer <- terra::crs(classes, proj = T)
    gisBase <- '/Applications/GRASS-7.8.app/Contents/Resources'
    initGRASS(gisBase = gisBase,
              home = tempdir(),
              gisDbase = tempdir(),  
              mapset = 'PERMANENT', 
              location = glue('mask_{tile_id}_{tile_index}'), 
              override = TRUE)
    execGRASS("g.proj", flags = "c", 
              proj4 = crs_mer)
    execGRASS('r.in.gdal', flags = c("o", "overwrite"),
              input = here(
                  glue('{refine_dir}/{tile_id}_{tile_index}_label.tif')),
              band = 1,
              output = "label")
    execGRASS("g.region", raster = "label")
    
    use_sf()
    if (nrow(mask_vct) > 0){
        writeVECT(mask_vct, 'mask', v.in.ogr_flags = 'overwrite')
        execGRASS('v.to.rast', flags = c("overwrite"),
                  parameters = list(input = 'mask', 
                                    output = 'mask',
                                    use = 'attr',
                                    attribute_column = 'landcover'))
        mask_path <- tempfile()
        execGRASS('r.out.gdal', flags = c("m", "overwrite"),
                  output = mask_path,
                  input = "mask")
        mask <- rast(mask_path)
        classes <- classes * is.na(mask)
        classes <- cover(classes, mask, values = 0)
        rm(mask)
    }; rm(mask_vct)
    
    if (apply_line_mask){
        mask_line <- read_sf(here('results/north/mask_line.geojson'))
        crs_mer <- crs(classes)
        mask_line <- mask_line %>% 
            st_make_valid() %>% 
            st_intersection(tile_sf) %>%
            st_transform(crs = crs_mer) %>% 
            dplyr::select(landcover)
        
        writeVECT(mask_line, 'mask', v.in.ogr_flags = 'overwrite')
        execGRASS('v.to.rast', flags = c("overwrite"),
                  parameters = list(input = 'mask', 
                                    output = 'mask',
                                    use = 'attr',
                                    attribute_column = 'landcover'))
        mask_path <- tempfile()
        execGRASS('r.out.gdal', flags = c("m", "overwrite"),
                  output = mask_path,
                  input = "mask")
        mask <- rast(mask_path)
        classes <- classes * is.na(mask)
        classes <- cover(classes, mask, values = 0)
        rm(mask, mask_line)
    }
    
    # Save out
    names(classes) <- 'label'
    writeRaster(classes, 
                file.path(refine_dir, 
                          glue('{tile_id}_{tile_index}_label.tif')),
                overwrite = T,
                wopt = list(datatype = 'INT1U',
                            gdal=c("COMPRESS=LZW")))
}

# After all checks finished, run these lines to 
# all tiles
tiles_all <- here('results/north/catalog_sample_tiles.geojson') %>% 
    read_sf() %>% 
    mutate(name_temp = paste(tile, index, sep = '_'))

# Tiles get modified
refine_tiles <- list.files(here('results/north/refine_labels')) %>% 
    str_extract(., '[0-9]+-[0-9]+_[0-9]+')

# Tiles not get modified
guess_tiles <- list.files(here('results/north/guess_labels'),
                          pattern = 'guess') %>% 
    str_extract(., '[0-9]+-[0-9]+_[0-9]+')
guess_tiles <- setdiff(guess_tiles, refine_tiles)

# modify tiles
tiles_all <- tiles_all %>% 
    mutate(modify = ifelse(name_temp %in% refine_tiles, 'yes', 'no')) %>% 
    select(-name_temp)
st_write(tiles_all, 
         here('results/north/catalog_sample_tiles_update.geojson'))

# Copy not modified labels
lapply(guess_tiles, function(tile_nm){
    copy_from <- file.path(here('results/north/guess_labels'), 
                           glue('guess_{tile_nm}.tif'))
    copy_to <- file.path(here('results/north/refine_labels'), 
                         glue('{tile_nm}_label.tif'))
    file.copy(copy_from, copy_to)
})
